requirements/prod.txt
Django==4.2.5
django-ckeditor==6.7.0
django-cleanup==8.0.0
Pillow==10.1.0
python-dotenv==1.0.0
sorl-thumbnail==12.10.0
transliterate==1.10.2
